# Enable reader/writer re-use, reduce allocations, fix benchmarks