--- a/clz_asm.go
+++ b/clz_asm.go
@@ -1,5 +1,0 @@
-// +build amd64
-
-package fpc
-
-func clzBytes(val uint64) uint64