--- a/writer.go
+++ b/writer.go
@@ -53,6 +53,24 @@ func NewWriterLevel(w io.Writer, level int) (*Writer, error) {
 	return z, nil
 }
 
+func (w *Writer) Reset(writer io.Writer, level int) error {
+	if level < 1 || level > MaxCompression {
+		return fmt.Errorf("fpc: invalid compression level: %d", level)
+	}
+
+	// Need to flush out what we already have
+	if err := w.Flush(); err != nil {
+		return fmt.Errorf("fpc: error flushing existing writer: %v", err)
+	}
+
+	// Reset to the new writer
+	w.w = writer
+	w.level = level
+	w.wroteHeader = false
+	w.closed = false
+	return w.enc.reset(writer, uint(level))
+}
+
 // Write interprets b as a stream of byte-encoded, 64-bit IEEE 754
 // floating point values. The length of b must be a multiple of 8 in
 // order to match this expectation.