… clz_amd64.s
… clz_test.go